+++ lib/playgrounder.dart
@@ -8,6 +8,7 @@ export 'src/knobs/knob_relevance.dart';
 export 'src/knobs/scale_knob.dart';
 export 'src/knobs/step_knob.dart';
 export 'src/knobs/switch_knob.dart';
+export 'src/playground/inspector/playground_actions.dart';
 export 'src/playground/playground.dart';
 export 'src/playground/playground_action.dart';
 export 'src/playground/playground_preset.dart';